requirements.txt
Babel
jinja2
ruamel.yaml
future
pillow>=6
pycryptodomex
imagesize
tqdm
